Flask_backend_python_login/requirements.txt
fastapi
passlib
python-jose
pymongo
python-multipart
uvicorn
pydantic
bcrypt